refactor(shared): move page-transition to src/components/shared/
Step 2 of desktop/mobile UI separation refactor.
- git mv page-transition.tsx -> shared/page-transition.tsx
- Update imports in (public)/layout.tsx and mockups/layout.tsx

PageTransition has no breakpoint signal and is used identically by
both eventual desktop and mobile shells. The mockups/layout.tsx import
is also updated since the Step 1 thin wrapper still uses it.

Plan: ~/.claude/plans/name-codex-review-merry-thunder.md
